+++ packages/url_launcher/url_launcher/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 6.0.0-nullsafety.5
+
+* Document that the web plugin is not endorsed in the `nullsafety` prerelease for now.
+
 ## 6.0.0-nullsafety.4
 
 * Update the example app: remove the deprecated `RaisedButton` and `FlatButton` widgets.
@@ -17,6 +21,7 @@
 ## 6.0.0-nullsafety
 
 * Migrate to null safety.
+* **Breaking change**: web plugins aren't endorsed in null-safe plugins yet.
 
 ## 5.7.13